ajout sécurité supp artistes, musiques, playlists

+++ code/src/php/deleteArtist.php
@@ -1,11 +1,19 @@
-<?php
-require "template/header.php";
+<?php require "template/header.php";
 
 if(isLogged() && (isAdmin()))
-{   
-    $idArtist = $_GET["idArtist"];
-    $deleteArtist = $db->deleteOneArtist($idArtist);
-    header('Location: allartists.php');  
+{     
+    // Vérifie que le get n'est pas vite, vérifie si le get est bien numérqiue -> rejete le code html et php (+ sécurisé)
+    if(!isset($_GET['idArtist']) OR !is_numeric($_GET['idArtist']))
+    {
+        header('Location:404.php');
+    }
+    // Si tout est ok -> appelle les fonctions
+    else
+    {
+        $idArtist = $_GET['idArtist'];
+        $deleteArtist = $db->deleteOneArtist($idArtist);
+        header('Location: allartists.php');    
+    } 
 }
 else
 {
