+++ code/src/php/deleteMusic.php
@@ -1,12 +1,19 @@
-<?php
-
-require "template/header.php";
+<?php require "template/header.php";
 
 if(isLogged() && (isAdmin()))
 {
-    $idMusic = $_GET["idMusic"];
-    $deleteArtist = $db->deleteOneMusic($idMusic);
-    header('Location: alltitle.php');  
+    // Vérifie que le get n'est pas vite, vérifie si le get est bien numérqiue -> rejete le code html et php (+ sécurisé)
+    if(!isset($_GET['idMusic']) OR !is_numeric($_GET['idMusic']))
+    {
+        header('Location:404.php');
+    }
+    // Si tout est ok -> appelle les fonctions
+    else
+    {
+        $idMusic = $_GET['idMusic'];
+        $deleteArtist = $db->deleteOneMusic($idMusic);
+        header('Location: alltitle.php');   
+    } 
 }
 else
 {
